+++ src/app/router/RouterComponent.jsx
@@ -3,7 +3,7 @@ import { Route, Routes } from 'react-router-dom'
 import { Bookmark } from '@/pages/bookmark/ui/Bookmark'
 import { FontDetail } from '@/pages/detail/ui/FontDetail'
 import { Explorer } from '@/pages/explorer/ui/Explorer'
-import { Home } from '@/pages/home/ui/Home'
+import { Home } from '@/pages/home/Home'
 import { FontDesignProgress } from '@/pages/myFonts/ui/FontDesignProgress'
 import { MyCustomFont } from '@/pages/myFonts/ui/MyCustomFont'
 import { MyFonts } from '@/pages/myFonts/ui/MyFonts'
