+++ src/shared/ui/EmptyMessage.jsx
@@ -7,7 +7,7 @@ export const EmptyMessage = ({ message }) => {
 const Container = styled.p`
   ${({ theme }) => `
     ${theme.flexBox('row', 'center', 'center')}
-    ${theme.font(400, theme.colors.grey[600])}
+    ${theme.font(500, theme.colors.grey[600])}
   `}
   width: 100%;
   height: 12rem;
