+++ src/widgets/font-item/ui/FontItemWithButtons.jsx
